Break out our video and image templates into fragments that can be reused
These fragments will be how we build the resulting JobTemplate with modified values.

Because we need to support a variable number of compositions in the same JobTemplate we need to generate the resulting template.json.

These fragments allow us to visually see the individual components that will create the template.json while still allowing us to programmatically create it.
The hope is that new developers coming to this project still gain some ease of modifying these while the project gains the power to support multiple comps per submission.

Signed-off-by: Alex Hughes <ahughesalex@gmail.com>